+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Improved performance of guessing decimals
+
+
 ## [0.0.2] - 2019-08-30
 
 ### Fixed